DDFitPar class definition

DDFitPar class definition

Slots

FIT_V

numeric representing the value of the Fit

FIT_V_self

numeric representing the value of the Fit with the population (if specified)

FIT_N

numeric representing the number of evaluation points used in the fit

START_VALUE

character specifying the method used for the start value problem

METHOD

character specifying the method used for fitting

S_SAMPLING

logical if super sampling was used

TRIALS_TOAL

numeric specifying the total numbers of trials that were simulated

TRIALS_SAMPLE

numeric specifying the number of trials inside a given sample

PERF

character specifying the performance analysis error