DDModel class definition

DDModel class definition

Slots

ID

character that represents the name of the model to be used (i.e. "DSTP","DMC","SSP")

MM

list of matrix that contain values which map custom parameters to correspondend modelparameters

DM

matrix that contains the domain of all custom parameters (and grid size steps)

SP

matrix that contains a set of simulation-parameters important for simulation

RF

list of numeric vectors that contain the percentiles of the representation.