Function to fit a given DDRep to a given DDModel

Function to fit a given DDRep to a given DDModel

Fit_DDModel(
  model = NULL,
  data = NULL,
  DL_model = NULL,
  DL_scale = list(INPUT = list(Center = FALSE, Stddev = FALSE), OUTPUT = list(Center =
    FALSE, Stddev = FALSE)),
  grid_path = NULL,
  trials = 10000L,
  s_sampling = FALSE,
  simplex_struc = c(0L),
  simplex_coef = list(alpha = 1, beta = 0.5, gamma = 2, sigma = 0.5, tol = 1e-07, nfunc
    = 120L, nshrink = 3L)
)

Arguments

model
DDModel object

data
DDRep object or list of DDRep objects

DL_model
(optional) Model in the form of a keras neural network model

DL_scale
(optional) list containing mean and sd of the transformation used in the deep learning model while training.

grid_path
(optional) path to a directory containing a .GRID fileset. If NULL the model will be fitted using 20 randomly drawn startparametersets from the model-DOMAIN.

trials
integer indicating the number of trials used while fitting (s_sampling = FALSE) or the maximum number of trials used while super sampling (s_sampling = TRUE)

s_sampling
bool indicating super sampling while fitting

simplex_struc
numeric vector containing the number of simplex iterations per sorting cycle.

simplex_coef
list containing named parameters for the SIMPLEX

Value

DDFit object

Details

This is a rather complex function that is very flexible depending on your input!

Random Fitting:

Exclude all optional parameters

Grid Fitting:

Specify a grid_path

Deep Learning Fitting

Specify DL_model and DL_scale